スケジュールの一覧の部分を別コンポーネント化
